Move FileData to separate file
