+++ src/org/omegat/core/statistics/FileData.java
@@ -1,0 +1,40 @@
+/**************************************************************************
+ OmegaT - Computer Assisted Translation (CAT) tool
+          with fuzzy matching, translation memory, keyword search,
+          glossaries, and translation leveraging into updated projects.
+
+ Copyright (C) 2009 Alex Buloichik
+               2010 Didier Briel
+               2020 Aaron Madlon-Kay
+               Home page: http://www.omegat.org/
+               Support center: https://omegat.org/support
+
+ This file is part of OmegaT.
+
+ OmegaT is free software: you can redistribute it and/or modify
+ it under the terms of the GNU General Public License as published by
+ the Free Software Foundation, either version 3 of the License, or
+ (at your option) any later version.
+
+ OmegaT is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU General Public License for more details.
+
+ You should have received a copy of the GNU General Public License
+ along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ **************************************************************************/
+
+package org.omegat.core.statistics;
+
+public class FileData {
+    public String filename;
+    public StatCount total, unique, remaining, remainingUnique;
+
+    public FileData() {
+        total = new StatCount();
+        unique = new StatCount();
+        remaining = new StatCount();
+        remainingUnique = new StatCount();
+    }
+}
